integration tests for the new folder structure
