+++ vizro-core/tests/integration/conftest.py
@@ -1,0 +1,22 @@
+# ruff: noqa: F403, F405
+import os
+
+import chromedriver_autoinstaller_fix
+import pytest
+
+
+# Use monkeypatch as a session-scoped fixture.
+# Taken from https://github.com/pytest-dev/pytest/issues/363#issuecomment-1335631998.
+@pytest.fixture(scope="session")
+def monkeypatch_session():
+    with pytest.MonkeyPatch.context() as mp:
+        yield mp
+
+
+@pytest.fixture(scope="session", autouse=True)
+def setup_integration_test_environment(monkeypatch_session):
+    # Dash debug mode seems to interfere with the tests, so we disable it here. Note "false" as a string is correct.
+    monkeypatch_session.setenv("DASH_DEBUG", "false")
+    # We only need to install chromedriver outside CI.
+    if not os.getenv("CI"):
+        chromedriver_autoinstaller_fix.install()
