+++ vizro-core/tests/integration/test_examples_demo.py
@@ -1,29 +1,10 @@
 # ruff: noqa: F403, F405
-import os
 import runpy
 from pathlib import Path
 
-import chromedriver_autoinstaller_fix
 import pytest
 
 from vizro import Vizro
-
-
-# Use monkeypatch as a session-scoped fixture.
-# Taken from https://github.com/pytest-dev/pytest/issues/363#issuecomment-1335631998.
-@pytest.fixture(scope="session")
-def monkeypatch_session():
-    with pytest.MonkeyPatch.context() as mp:
-        yield mp
-
-
-@pytest.fixture(scope="session", autouse=True)
-def setup_integration_test_environment(monkeypatch_session):
-    # Dash debug mode seems to interfere with the tests, so we disable it here. Note "false" as a string is correct.
-    monkeypatch_session.setenv("DASH_DEBUG", "false")
-    # We only need to install chromedriver outside CI.
-    if not os.getenv("CI"):
-        chromedriver_autoinstaller_fix.install()
 
 
 @pytest.fixture(params=["default", "from_yaml"])
@@ -35,7 +16,7 @@ def demo_dashboard(request, monkeypatch):
 
 # Ignore deprecation warning until this is solved: https://github.com/plotly/dash/issues/2590
 @pytest.mark.filterwarnings("ignore:HTTPResponse.getheader()")
-def test_demo_dashboard(dash_duo, features_dashboard):
-    app = Vizro(assets_folder=Path(__file__).parents[2] / "examples/demo/assets").build(features_dashboard).dash
+def test_demo_dashboard(dash_duo, demo_dashboard):
+    app = Vizro(assets_folder=Path(__file__).parents[2] / "examples/demo/assets").build(demo_dashboard).dash
     dash_duo.start_server(app)
     assert dash_duo.get_logs() == []
